refactor(stylelint-copyright): migrate to TS
