+++ packages/stylelint-copyright/package.json
@@ -2,14 +2,19 @@
   "name": "stylelint-copyright",
   "version": "2.0.0-beta.20",
   "description": "Stylelint plugin to check CSS files for a copyright header.",
-  "main": "index.js",
+  "main": "lib/index.js",
   "license": "MIT",
+  "scripts": {
+    "build": "tsc",
+    "watch": "tsc --watch"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/facebook/docusaurus.git",
     "directory": "packages/stylelint-copyright"
   },
   "dependencies": {
-    "stylelint": "^14.8.2"
+    "stylelint": "^14.8.2",
+    "tslib": "^2.4.0"
   }
 }
